--- src/library-authoring/common/context.js
@@ -1,10 +1,0 @@
-// @ts-check
-import React from 'react';
-
-const LibraryContext = React.createContext({
-  sidebarBodyComponent: /** @type{null|string} */ (null),
-  closeLibrarySidebar: /** @type{function} */ (() => {}),
-  openAddContentSidebar: /** @type{function} */ (() => {}),
-});
-
-export default LibraryContext;
