--- src/library-authoring/common/index.js
@@ -1,3 +1,0 @@
-// @ts-check
-export { default as LibraryContext } from './context';
-export { default as useLibraryContext } from './hooks';
